refactor: move error handling code to error and debug
