+++ src/routes/middleware/ErrorHandler.tsx
@@ -1,0 +1,25 @@
+import express from 'express';
+import { sendError } from '../../lib/error/sendError';
+import { UploadedFile } from '../../lib/storage/types';
+import { isLimitError } from '../../lib/misc/isLimitError';
+import { isEmptyPayload } from '../../lib/misc/isEmptyPayload';
+import { perserveFilesForDebugging } from '../../lib/debug/perserveFilesForDebugging';
+
+export default function ErrorHandler(
+  res: express.Response,
+  req: express.Request,
+  err: Error
+) {
+  const uploadedFiles = req.files as UploadedFile[];
+  const skipError = isLimitError(err);
+
+  if (!skipError) {
+    sendError(err);
+    if (!isEmptyPayload(uploadedFiles)) {
+      perserveFilesForDebugging(req.files as UploadedFile[], err);
+    }
+  }
+
+  res.set('Content-Type', 'text/plain');
+  res.status(400).send(err.message);
+}
